requirements.txt
dedalus_labs>=0.1.0
anthropic>=0.40.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
